+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `InfoLines` component.
 - Fix: Add `width` attribute on payment button svg tag (fixes firefox bug).
 - Feature: Add `color-error` color text utility.
 - Feature: For `xs` media-query, `FormActions` component is placed in columns.